requirements.txt
alembic==1.16.1
fastapi==0.115.12
psycopg2-binary==2.9.10
python-dotenv==1.0.0
SQLAlchemy==2.0.41
uvicorn==0.34.2